refactor: Rename `lib/plugins/aws/deploy/lib/cleanupS3Bucket.js`
