+++ lib/plugins/aws/deploy/index.js
@@ -11,7 +11,7 @@ const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
 const getExecuteChangeSetParams = require('../lib/get-execute-change-set-params');
 const waitForChangeSetCreation = require('../lib/wait-for-change-set-creation');
 const createStack = require('./lib/create-stack');
-const cleanupS3Bucket = require('./lib/cleanupS3Bucket');
+const cleanupS3Bucket = require('./lib/cleanup-s3-bucket');
 const uploadArtifacts = require('./lib/uploadArtifacts');
 const validateTemplate = require('./lib/validateTemplate');
 const updateStack = require('../lib/update-stack');
