+++ lib/plugins/aws/deploy/lib/cleanup-s3-bucket.js
