+++ test/unit/lib/plugins/aws/deploy/lib/cleanup-s3-bucket.test.js
